refactor: auth 파일의 getRefreshToken 함수를 interceptor 파일로 이동
